+++ tests/test_cli.py
@@ -1,14 +1,14 @@
 """md2pdf tests for the CLI."""
 
 import json
+from importlib import metadata
 
 import pytest
-from click.testing import CliRunner
 
 from md2pdf.cli import cli, parse_config
 from md2pdf.exceptions import ValidationError
 
-from .defaults import INPUT_CSS, INPUT_MD, OUTPUT_PDF
+from .defaults import DEFAULT_OUTPUT_PDF, INPUT_CSS, INPUT_MD, OUTPUT_PDF
 
 
 @pytest.mark.parametrize(
@@ -50,80 +50,106 @@ def test_parse_config_invalid_config(config):
         parse_config(config)
 
 
-def test_print_usage_when_no_args():
+def test_print_usage_when_no_args(cli_runner):
     """Print usage when no arguments are passed."""
-    runner = CliRunner()
-    result = runner.invoke(cli)
-    expected = "Usage: md2pdf [OPTIONS] MD PDF"
+    result = cli_runner.invoke(cli)
+    expected = "Usage: main [OPTIONS]"
     assert result.exit_code == 2
     assert expected in result.output
 
 
-def test_print_usage_when_partial_args():
-    """Print usage when required arguments are missing."""
-    runner = CliRunner()
-    result = runner.invoke(cli, ["input.md"])
-    expected = "Usage: md2pdf [OPTIONS] MD PDF"
+def test_exit_when_no_markdown_input(cli_runner):
+    """Exit with an error message when not input markdown files are passed."""
+    result = cli_runner.invoke(cli, ["-o", "test.pdf"])
+    expected = "No markdown input file. See `--help`"
     assert result.exit_code == 2
     assert expected in result.output
 
 
-def test_raise_IOError_when_markdown_input_file_does_not_exists():
-    """Raise an I/O error when markdown input file does not exist."""
-    runner = CliRunner()
-    result = runner.invoke(cli, ["input.md", "output.pdf"])
-    expected = "Error: Invalid value for 'MD': Path 'input.md' does not exist."
-    assert result.exit_code == 2
-    assert expected in result.output
-
-
-def test_raise_IOError_when_stylesheet_does_not_exists():
-    """Raise an I/O error when CSS input file does not exist."""
-    runner = CliRunner()
-    result = runner.invoke(cli, ["--css=styles.css", str(INPUT_MD), str(OUTPUT_PDF)])
-    expected = (
-        "Error: Invalid value for '--css' / '-c': Path 'styles.css' does not exist."
+def test_exit_when_multiple_markdown_and_pdf_options(cli_runner):
+    """Exit with an error message when called with multitple input & output options."""
+    result = cli_runner.invoke(
+        cli, ["-i", str(INPUT_MD), "-i", str(INPUT_MD), "-o", "test.pdf"]
     )
+    expected = "PDF output option `--output/-o` cannot be used with multiple input."
     assert result.exit_code == 2
     assert expected in result.output
 
 
-def test_generate_pdf_from_markdown_source_file():
+def test_command_version(cli_runner):
+    """Print program version when asked."""
+    result = cli_runner.invoke(cli, "-V")
+    expected = metadata.version("md2pdf")
+    assert result.exit_code == 0
+    assert expected in result.output
+
+
+def test_command_raise_error_when_markdown_input_file_does_not_exists(cli_runner):
+    """Raise an I/O error when markdown input file does not exist."""
+    result = cli_runner.invoke(cli, ["-i", "input.md"])
+    expected = "Invalid value for '--input' / '-i': Path 'input.md' does not exist."
+    assert result.exit_code == 2
+    assert expected in result.output
+
+
+def test_command_raise_error_when_stylesheet_does_not_exists(cli_runner):
+    """Raise an I/O error when CSS input file does not exist."""
+    result = cli_runner.invoke(
+        cli, ["-c", "styles.css", "-i", str(INPUT_MD), "-o", str(OUTPUT_PDF)]
+    )
+    expected = "Invalid value for '--css' / '-c': Path 'styles.css' does not exist."
+    assert result.exit_code == 2
+    assert expected in result.output
+
+
+def test_generate_pdf_from_markdown_source_file(cli_runner):
     """Generate a PDF from a markdown file."""
     assert not OUTPUT_PDF.exists()
-    runner = CliRunner()
-    runner.invoke(cli, [str(INPUT_MD), str(OUTPUT_PDF)])
+    result = cli_runner.invoke(cli, ["-i", str(INPUT_MD), "-o", str(OUTPUT_PDF)])
+    assert result.exit_code == 0
     assert OUTPUT_PDF.exists()
 
 
-def test_generate_pdf_from_markdown_source_file_and_stylesheet():
+def test_generate_pdf_from_markdown_source_file_without_output(cli_runner):
+    """Generate a PDF from a markdown file with no output argument."""
+    assert not DEFAULT_OUTPUT_PDF.exists()
+    result = cli_runner.invoke(cli, ["-i", str(INPUT_MD)])
+    assert result.exit_code == 0
+    assert DEFAULT_OUTPUT_PDF.exists()
+
+
+def test_generate_pdf_from_markdown_source_file_and_stylesheet(cli_runner):
     """Generate a PDF from a markdown and a CSS file."""
     assert not OUTPUT_PDF.exists()
-    runner = CliRunner()
-    runner.invoke(cli, [f"--css={INPUT_CSS}", str(INPUT_MD), str(OUTPUT_PDF)])
+    result = cli_runner.invoke(
+        cli, ["-c", str(INPUT_CSS), "-i", str(INPUT_MD), "-o", str(OUTPUT_PDF)]
+    )
+    assert result.exit_code == 0
     assert OUTPUT_PDF.exists()
 
 
-def test_generate_pdf_with_an_extension():
+def test_generate_pdf_with_an_extension(cli_runner):
     """Generate a PDF from a markdown file with an extra extension."""
     assert not OUTPUT_PDF.exists()
-    runner = CliRunner()
-    runner.invoke(cli, ["--extras", "footnotes", str(INPUT_MD), str(OUTPUT_PDF)])
+    cli_runner.invoke(
+        cli, ["--extras", "footnotes", "-i", str(INPUT_MD), "-o", str(OUTPUT_PDF)]
+    )
     assert OUTPUT_PDF.exists()
 
 
-def test_generate_pdf_with_a_configured_extension():
+def test_generate_pdf_with_a_configured_extension(cli_runner):
     """Generate a PDF from a markdown file with an extra configured extension."""
     assert not OUTPUT_PDF.exists()
-    runner = CliRunner()
-    runner.invoke(
+    cli_runner.invoke(
         cli,
         [
-            "--extras",
+            "-e",
             "footnotes",
-            "--config",
+            "-C",
             json.dumps({"footnotes": {"BACKLINK_TEXT": "link"}}),
+            "-i",
             str(INPUT_MD),
+            "-o",
             str(OUTPUT_PDF),
         ],
     )